Refactor testing utilities and remove unused exports
- Removed unused testing utilities from `index.ts` and `hook-testing-utils.ts`.
- Centralized the `createTestWrapper` function in `test-helpers.ts` for better organization.
- Updated unit tests to use hardcoded mock data for consistency.
