+++ apps/admin-x-framework/src/index.ts
@@ -18,20 +18,3 @@ export {Link, NavLink, Outlet, useLocation, useParams, useSearchParams, redirect
 // Data fetching
 export type {InfiniteData} from '@tanstack/react-query';
 export {useQueryClient} from '@tanstack/react-query';
-
-// Testing utilities
-export type {MockApiHookConfig} from './test/hook-testing-utils';
-export {
-    createMockApiReturn,
-    mockApiHook,
-    createMockApiHook,
-    createLoadingMock,
-    createErrorMock,
-    createSuccessMock,
-    mockDataFactories,
-    waitForMockCall,
-    waitForMockCalls,
-    resetAllMocks,
-    createTypedMockApiHook,
-    createTestWrapper
-} from './test/hook-testing-utils';
