+++ apps/admin-x-settings/package.json
@@ -28,7 +28,7 @@
     "lint": "yarn run lint:js",
     "lint:js": "eslint --ext .js,.ts,.cjs,.tsx --cache src test",
     "test:unit": "vitest run --config vitest.config.ts",
-    "test:acceptance": "NODE_OPTIONS='--experimental-specifier-resolution=node --no-warnings' VITE_TEST=true playwright test",
+    "test:acceptance": "NODE_OPTIONS='--experimental-specifier-resolution=node --no-warnings' playwright test",
     "test:acceptance:slowmo": "TIMEOUT=100000 PLAYWRIGHT_SLOWMO=100 yarn test:acceptance --headed",
     "test:acceptance:full": "ALL_BROWSERS=1 yarn test:acceptance",
     "preview": "vite preview"
@@ -72,7 +72,6 @@
       },
       "test:acceptance": {
         "dependsOn": [
-          "test:acceptance",
           "^build"
         ]
       }
