+++ apps/stats/test/unit/hooks/useFeatureFlag.test.tsx
@@ -25,6 +25,7 @@ describe('useFeatureFlag', () => {
 
     it('returns loading state when data is loading', () => {
         mocks.mockUseGlobalData.mockReturnValue({
+            ...mocks.mockUseGlobalData.mock.results[0]?.value || {},
             isLoading: true,
             settings: []
         });
