+++ apps/stats/test/utils/test-helpers.ts
@@ -1,18 +1,36 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import React from 'react';
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import {
     createErrorMock,
     createLoadingMock,
     createMockApiReturn,
     createSuccessMock,
-    createTestWrapper,
     mockApiHook,
     resetAllMocks
-} from '../../../admin-x-framework/src/test/hook-testing-utils';
+} from '@tryghost/admin-x-framework/test/hook-testing-utils';
 import {responseFixtures} from '@tryghost/admin-x-framework/test/acceptance';
 import {vi} from 'vitest';
 
 // Import types from API modules
 import type {NewsletterStatsResponseType, TopPostsStatsResponseType} from '@tryghost/admin-x-framework/api/stats';
+
+// Create a test wrapper with QueryClient
+export const createTestWrapper = () => {
+    const queryClient = new QueryClient({
+        defaultOptions: {
+            queries: {retry: false},
+            mutations: {retry: false}
+        }
+    });
+    
+    const Wrapper = ({children}: {children: React.ReactNode}) => (
+        React.createElement(QueryClientProvider, {client: queryClient}, children)
+    );
+    Wrapper.displayName = 'TestWrapper';
+    
+    return Wrapper;
+};
 
 // Re-export centralized utilities for convenience
 export {
@@ -20,7 +38,6 @@ export {
     createLoadingMock,
     createMockApiReturn,
     createSuccessMock,
-    createTestWrapper,
     mockApiHook,
     resetAllMocks
 };
